refs #203: Renamed the textbox content file.
